refactor directory and files structure in app-reports component
