--- app-reports/src/main/kotlin/com/usvision/reports/PlanExecutioner.kt
@@ -1,7 +1,0 @@
-package com.usvision.reports
-
-import com.usvision.model.System
-
-interface PlanExecutioner {
-    fun execute(plan: Plan, system: System): Report
-}
