--- app-reports/src/main/kotlin/com/usvision/reports/Planner.kt
@@ -1,5 +1,0 @@
-package com.usvision.reports
-
-interface Planner {
-    fun plan(reportRequest: ReportRequest): Plan
-}
