--- app-reports/src/main/kotlin/com/usvision/reports/ReportRequestGenerationException.kt
@@ -1,13 +1,0 @@
-package com.usvision.reports
-
-open class ReportRequestGenerationException(
-    message: String
-) : RuntimeException(message)
-
-class DetectorNotFoundException(
-    name: String
-) : ReportRequestGenerationException("Detector $name not found")
-
-class ClassIsNotDetectorException(
-    name: String
-) : ReportRequestGenerationException("Class $name is not Detector")
