+++ app-reports/src/main/kotlin/com/usvision/reports/exceptions/ClassIsNotDetectorException.kt
@@ -1,0 +1,5 @@
+package com.usvision.reports.exceptions
+
+class ClassIsNotDetectorException(
+    name: String
+) : ReportRequestGenerationException("Class $name is not Detector")
