+++ app-reports/src/main/kotlin/com/usvision/reports/exceptions/DetectorNotFoundException.kt
@@ -1,0 +1,5 @@
+package com.usvision.reports.exceptions
+
+class DetectorNotFoundException(
+    name: String
+) : ReportRequestGenerationException("Detector $name not found")
