+++ app-reports/src/main/kotlin/com/usvision/reports/exceptions/ReportRequestGenerationException.kt
@@ -1,0 +1,6 @@
+package com.usvision.reports.exceptions
+
+open class ReportRequestGenerationException(
+    message: String
+) : RuntimeException(message)
+
