+++ app-reports/src/main/kotlin/com/usvision/reports/exceptions/UnknownPresetException.kt
@@ -1,4 +1,4 @@
-package com.usvision.reports
+package com.usvision.reports.exceptions
 
 class UnknownPresetException(name: String)
     : RuntimeException("Preset $name is unknown")
