+++ app-reports/src/main/kotlin/com/usvision/reports/executioner/PlanExecutioner.kt
@@ -1,0 +1,9 @@
+package com.usvision.reports.executioner
+
+import com.usvision.model.System
+import com.usvision.reports.utils.Plan
+import com.usvision.reports.utils.Report
+
+interface PlanExecutioner {
+    fun execute(plan: Plan, system: System): Report
+}
