+++ app-reports/src/main/kotlin/com/usvision/reports/planner/Planner.kt
@@ -1,0 +1,8 @@
+package com.usvision.reports.planner
+
+import com.usvision.reports.utils.Plan
+import com.usvision.reports.utils.ReportRequest
+
+interface Planner {
+    fun plan(reportRequest: ReportRequest): Plan
+}
