+++ app-reports/src/main/kotlin/com/usvision/reports/utils/Report.kt
@@ -1,0 +1,8 @@
+package com.usvision.reports.utils
+
+import com.usvision.analyses.ArchitectureInsight
+import kotlin.reflect.KClass
+
+class Report(
+    private val value: Map<KClass<out ArchitectureInsight>, Any>
+) : Map<KClass<out ArchitectureInsight>, Any> by value
