+++ app-reports/src/main/kotlin/com/usvision/reports/utils/ReportRequest.kt
@@ -1,0 +1,8 @@
+package com.usvision.reports.utils
+
+import com.usvision.analyses.Detector
+import kotlin.reflect.KClass
+
+data class ReportRequest(
+    val detectors: Set<KClass<out Detector>>
+)
